requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-dotenv
django-filter
django-filter
djangorestframework-simplejwt
coverage